refactor: 전체적인 패키징 변경
