+++ src/main/java/lotto/constants/LottoConstant.java
@@ -1,7 +1,5 @@
 package lotto.constants;
 
-import lotto.domain.Lotto;
-
 public class LottoConstant {
     public static final int MIN_LOTTO_NUMBER = 1;
     public static final int MAX_LOTTO_NUMBER = 45;
