+++ src/main/java/lotto/domain/model/Lotto.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.domain.model;
 
 import java.util.HashSet;
 import java.util.List;
