+++ src/main/java/lotto/domain/model/LottoNumber.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.domain.model;
 
 import lotto.constants.ErrorMessage;
 import lotto.constants.LottoConstant;
