+++ src/test/java/lotto/domain/model/PurchaseMoneyTest.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.domain.model;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
